[Installer] Check if host files exist only then copy and link them

+++ install.sh
@@ -18,18 +18,20 @@ fi
 git clone https://github.com/tuxinaut/nix-home.git $INSTALL_PATH
 cd $INSTALL_PATH
 
-if [[ ! -e ./hosts/$HOST.nix ]]; then
+if [[ ! -e ./hosts/$HOST.nix && -e $SOURCE_PATH/configuration.nix ]]; then
   cp -v -- $SOURCE_PATH/configuration.nix ./hosts/$HOST.nix
 fi
 
-if [[ ! -e ./hosts/$HOST-hardware-configuration.nix ]]; then
+if [[ ! -e ./hosts/$HOST-hardware-configuration.nix && -e $SOURCE_PATH/hardware-configuration.nix ]]; then
   cp -v -- $SOURCE_PATH/hardware-configuration.nix ./hosts/$HOST-hardware-configuration.nix
   # TODO
   # SED for setting the right hardware path
   # Or do it in the right way via nix dsl
 fi
 
-ln -s ./hosts/$HOST.nix configuration.nix
+if [[ -e ./hosts/$HOST.nix ]]; then
+  ln -s ./hosts/$HOST.nix configuration.nix
+fi
 
 ln -s "." "${HOME}/.config/nixpkgs"
